fix: decouple react and react-dom imports from utils and make it tree-shakeable
